refactor: combine with other settings
